requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.47.2
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
passlib==1.7.4
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.47.2
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
passlib==1.7.4
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
